pyproject.toml
[tool.poetry]
name = "ragsearch"
version = "0.1.1"
description = "is a Python library designed for building a Retrieval-Augmented Generation (RAG) application that enables natural language querying over structured data. This tool leverages embedding models and a vector database (FAISS) to provide an efficient and scalable search engine."
authors = ["Mrutunjay Kinagi < mrutunjay.kinagi @ gmail.com >"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/mrutunjay-kinagi/ragsearch"
packages = [
    { include = "ragsearch", from = "libs" },
]

[tool.poetry.dependencies]
python = "^3.9"
pytest = "^8.3.3"
sphinx = "7.4.7"
sphinx-rtd-theme = "^3.0.2"
pylint = "^3.3.1"
flask = "^3.1.0"
cohere = "^5.13.3"
faiss-cpu = "^1.9.0.post1"
pandas = "^2.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
